created a submission

+++ script.js
@@ -11,7 +11,7 @@
           currentIndex++;
           //setTimeout(typeNextCharacter, 500 / parseInt(speedInput.value));
 			if(speedInput.value < 5 ){
-				setTimeout(typeNextCharacter, 50000 / parseInt(speedInput.value));
+				setTimeout(typeNextCharacter, 50000 / parseFloat(speedInput.value));
 			}
 	          else{
 				  setTimeout(typeNextCharacter, 5000 / parseInt(speedInput.value));
